requirements.txt
# -------------------------------
# Core Frameworks
# -------------------------------
fastapi==0.104.1
streamlit==1.28.1
uvicorn[standard]==0.24.0

# -------------------------------
# LangChain & AI Integration
# -------------------------------
langchain==0.2.17
langchain-community==0.2.19
langchain-openai==0.1.23
langchain-core==0.2.43
langchain-text-splitters==0.2.4
openai>=1.40.0,<2.0.0
tiktoken>=0.7,<1.0

# -------------------------------
# Vector Store & Embeddings
# -------------------------------
faiss-cpu==1.7.4
pickle-mixin==1.0.2

# -------------------------------
# PDF Processing
# -------------------------------
PyMuPDF==1.23.8
pdfplumber==0.10.3

# -------------------------------
# Environment & Validation
# -------------------------------
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# -------------------------------
# Networking & Requests
# -------------------------------
requests==2.31.0
httpx==0.25.2

# -------------------------------
# Data Processing
# -------------------------------
numpy==1.24.4
pandas==2.1.4

# -------------------------------
# Logging & Monitoring
# -------------------------------
python-json-logger==2.0.7
